[x86/Linux] (Partially) Enable FEATURE_EH_FUNCLETS
